+++ sdk-node/test/conformance/dispatch-contacts.ts
@@ -83,6 +83,8 @@ export function dispatchContacts(
     case 'activate':
     case 'cancel':
     case 'cancelImmediately':
+    case 'compNextCycle':
+    case 'uncompNextCycle':
     case 'markUnpaid':
     case 'bill':
     case 'renew':